refactor(test): remove testClone for all scales
