+++ __tests__/unit/scales/base.spec.ts
@@ -1,6 +1,5 @@
 import { Base, DEFAULT_OPTIONS } from '../../../src/scales/base';
 import { BaseOptions, Primitive } from '../../../src/types';
-import testClone from '../../utils/clone';
 import { ticks } from '../../../src/tick-method/basic';
 
 class Scale extends Base<BaseOptions> {
@@ -90,5 +89,30 @@ describe('Scale', () => {
     expect(mockFn).toBeCalled();
   });
 
-  testClone(Scale);
+  test('clone() return a scale belong to same class', () => {
+    const s = new Scale();
+    const s1 = s.clone();
+    expect(s1).toBeInstanceOf(Scale);
+  });
+
+  test('clone() return a scale with the same options as the original one', () => {
+    const s = new Scale();
+    const s1 = s.clone();
+    expect(s.getOptions()).toEqual(s1.getOptions());
+  });
+
+  test('clone() return a scale isolating change with the original one', () => {
+    const s = new Scale();
+    const s1 = s.clone();
+
+    s.update({
+      tickCount: 20,
+    });
+    expect(s1.getOptions().tickCount).toBe(5);
+
+    s1.update({
+      tickCount: 30,
+    });
+    expect(s.getOptions().tickCount).toBe(20);
+  });
 });
